fixed camera preview size, firebase_auth persists through firebase_auth demo screen
