+++ lib/camera/display_picture_screen.dart
@@ -12,7 +12,7 @@ class DisplayPictureScreen extends StatelessWidget {
       appBar: AppBar(title: Text('Display the Picture')),
       // The image is stored as a file on the device. Use the `Image.file`
       // constructor with the given path to display the image.
-      body: Image.file(File(imagePath)),
+      body: Center(child: Image.file(File(imagePath))),
     );
   }
 }
