+++ lib/firebase_auth/firebase_auth_page.dart
@@ -1,12 +1,39 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_sandbox/auth.dart';
 import 'package:flutter_sandbox/firebase_auth/Components/rounded_button.dart';
 import 'package:flutter_sandbox/firebase_auth/firebase_auth_login_page.dart';
 import 'package:flutter_sandbox/firebase_auth/firebase_auth_register_page.dart';
 import 'package:flutter_sandbox/screen_arguments.dart';
+import 'package:provider/provider.dart';
 
-class FirebaseAuthPage extends StatelessWidget {
+class FirebaseAuthLandingPage extends StatelessWidget {
   static const id = 'firebase_auth_page';
 
+  @override
+  Widget build(BuildContext context) {
+    return ChangeNotifierProvider(
+      create: (context) => Auth(),
+      child: LandingPage(),
+    );
+  }
+}
+
+class LandingPage extends StatelessWidget {
+  final _auth = FirebaseAuth.instance;
+
+  @override
+  Widget build(BuildContext context) {
+    print(_auth.currentUser);
+    if (_auth.currentUser == null) {
+      return FirebaseAuthLoggedOutState();
+    } else {
+      return FirebaseAuthLoggedInState();
+    }
+  }
+}
+
+class FirebaseAuthLoggedOutState extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final FirebaseAuthPageArgs args = ModalRoute.of(context).settings.arguments;
@@ -46,5 +73,27 @@ class FirebaseAuthPage extends StatelessWidget {
         ),
       ),
     );
+  }
+}
+
+class FirebaseAuthLoggedInState extends StatelessWidget {
+  final _auth = FirebaseAuth.instance;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: Text('Firebase Auth'),
+        ),
+        body: Center(
+          child: RoundedButton(
+            title: 'Sign out',
+            colour: Colors.lightBlueAccent,
+            onPressed: () {
+              _auth.signOut();
+              Navigator.pop(context);
+            },
+          ),
+        ));
   }
 }
